Move dtypes  jumpy.numpy module, update tests
